Now saving pers_periodograms and thresh_pks to pickled search object, not just csvs. Also fixed 0-planet initialization with k1.value call.

+++ rvsearch/periodogram.py
@@ -170,7 +170,7 @@ class Periodogram:
 
             for i, per in enumerate(per_array):
                 if self.verbose:
-                    limit = self.num_pers/float(self.workers)
+                    limit = int(self.num_pers/float(self.workers))
                     print(' {}'.format(i), '/', limit, end='\r')
                 # Reset posterior parameters to default values.
                 for k in self.default_pdict.keys():
